feat: Add Room and Hilt dependencies with create, update, and delete tasks
This commit introduces Room for local database management and Hilt for dependency injection. It includes the implementation of `TaskEntity`, `TaskDao`, and `ToDoDataBase` for data persistence. Additionally, it provides `AddTasksUseCase`, `UpdateTasksUseCase`, and `DeleteTasksUseCase` for handling task operations and `TaskRepository` to centralize data access. This commit also includes adding dependencies,  and refactoring `TasksViewModel` to manage the task state and interact with the use cases. Also implemented `GetTasksUseCase` to get all task of database. Finally, it implements `TaskUiState` for managing state of the screen.